Case z to signed int for comparison to suppress warning.

+++ include/prng_engine.h
@@ -229,7 +229,7 @@ namespace sitmo
 			void discard(uint64_t z)
 			{
 				// check if we stay in the current block
-				if (z < 8 - _o_counter)
+				if ((int64_t) z < 8 - _o_counter)
 				{
 					_o_counter += static_cast<unsigned short>(z);
 					return;
